[example] complete relay example with http and rtsp resouircesd behind the firewall
