+++ eyepop/worker/worker_endpoint.py
@@ -345,6 +345,8 @@ class WorkerEndpoint(Endpoint, WorkerClientSession):
             stream: BinaryIO | AsyncIterable[bytes] | Iterable[bytes],
             mime_type: str,
             video_mode: VideoMode | None = None,
+            is_live: bool | None = None,
+            captured_at_offset_ns: int | None = None,
             params: list[ComponentParams] | None = None,
             motion_detect: MotionDetectConfig | None = None,
             roi: Area | None = None,
@@ -356,6 +358,8 @@ class WorkerEndpoint(Endpoint, WorkerClientSession):
             stream=stream,
             mime_type=mime_type,
             video_mode=video_mode,
+            is_live=is_live,
+            captured_at_offset_ns=captured_at_offset_ns,
             component_params=params,
             motion_detect=motion_detect,
             roi=roi,
